refactor(tools): restructuring (wip)
